This commit will add login logic

+++ js/login.js
@@ -3,13 +3,24 @@
 
 
  // Name and Password from the register-form
-var name = document.getElementById('name');
+var name = document.getElementById('nme');
 var pw = document.getElementById('pw');
+var pwConfirm = document.getElementById('pwc')
+
 
 // storing input from register-form
 function store() {
-    localStorage.setItem('name', nme.value);
-    localStorage.setItem('pw', pw.value);
+    if (pw.value === pwc.value && pwConfirm !=null){
+        localStorage.setItem('name', nme.value);
+        localStorage.setItem('pw', pw.value);
+        alert("registration Successful please go back to login")
+    }
+    else {
+        alert("Mismatching passwords")
+    }
+     
+
+
 }
 
 // user  Authentication 
@@ -25,14 +36,32 @@ function check(){
     var userPw = document.getElementById('userPw')
     // check if stored data from register-form is equal to data from login form
     if(userName.value == storedName && userPw.value == storedPw) {
-        alert('Success.');
+        alert('Success.')
+        window.open("http://mywebsite.com/home.html");
     }else {
         alert('Invalid Username/ password');
     }
 
-    let btn = document.getElementById("btn");
+    
+}
+let btn = document.getElementById("btn");
     btn.addEventListener('click', event => {
       test();
     });
+let btnConfirm = document.getElementById("confirm");
+    btnConfirm.addEventListener('click', event => {
+      store();
+    });
    
-}
+// user registration section
+$(document).ready(function(){
+    $("button#register").click(function(){
+        $(".layout").hide();
+        $(".reg").show();
+    })
+    $("button#back").click(function(){
+        $(".layout").show();
+        $(".reg").hide();
+    })
+    
+})
